+++ .env
@@ -7,3 +7,4 @@ PGPORT=5432
 ADMIN_USERNAME=adminuser077
 ADMIN_EMAIL=admin@example.com
 ADMIN_PASSWORD=goat96321478
+VITE_API_BASE_URL=http://localhost:3000
