+++ capstonFE/src/components/Login.jsx
@@ -16,7 +16,8 @@ function Login() {
   const navigate = useNavigate();
   const location = useLocation();
 
-  const API_BASE_URL = "http://localhost:3000"; 
+   const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || "http://localhost:3000"; 
+
 
   // Determine initial mode based on current path
   const [isLoginMode, setIsLoginMode] = useState(
