+++ capstoneBE/package.json
@@ -9,7 +9,7 @@
     "seed": "node server/db/seed.js",
     "seed:dev": "nodemon server/db/seed.js",
     "create-admin": "node server/scripts/createAdmin.js"
-    },
+  },
   "keywords": [],
   "author": "",
   "license": "ISC",
