+++ capstoneBE/server/api/reviewRoutes.js
@@ -1,4 +1,3 @@
-
 const express = require('express');
 const router = express.Router();
 const {
@@ -8,18 +7,13 @@ const {
   fetchReviews,
   deleteReview,
   updateReview,
+  getReviewById,
 } = require('../db/review.js');
 const { authenticateToken, adminAuth } = require('./middlewares.js');
 
-// base route and return for the api for reviews
-
-// /api/reviews
-
 // create a review for an album
 router.post('/album/:albumId/create', async (req, res, next) => {
-  console.log('route logic');
   try {
-    console.log(req.body);
     const review = await createReview(
       req.params.albumId,
       req.body.userId,
@@ -28,7 +22,6 @@ router.post('/album/:albumId/create', async (req, res, next) => {
       req.body.rating,
       req.body.favorite
     );
-
     res.status(201).send(review);
   } catch (error) {
     next(error);
@@ -66,7 +59,7 @@ router.get('/user/:userId/', async (req, res, next) => {
 });
 
 // delete a review by id (admin only)
-router.delete('/:id/delete', authenticateToken, adminAuth, async (req, res, next) => {
+router.delete('/admin/:id/delete', authenticateToken, adminAuth, async (req, res, next) => {
   try {
     const response = await deleteReview(req.params.id);
     res.status(200).json({ message: 'Review deleted successfully', review: response });
@@ -75,9 +68,37 @@ router.delete('/:id/delete', authenticateToken, adminAuth, async (req, res, next
   }
 });
 
-// update a review by id
-router.put('/:id/update', async (req, res, next) => {
+// delete a review by id (non-admin)
+router.delete('/:id/delete', authenticateToken, async (req, res, next) => {
   try {
+    const review = await getReviewById(req.params.id);
+    if (review.user_id !== req.user.id) {
+      return res.status(403).json({ message: 'You are not authorized to delete this review' });
+    }
+    const response = await deleteReview(req.params.id);
+    res.status(200).json({ message: 'Review deleted successfully', review: response });
+  } catch (error) {
+    next(error);
+  }
+});
+
+// update a review by id (admin only)
+router.put('/admin/:id/update', authenticateToken, adminAuth, async (req, res, next) => {
+  try {
+    const response = await updateReview(req.params.id, req.body.review);
+    res.send(response);
+  } catch (error) {
+    next(error);
+  }
+});
+
+// update a review by id (non-admin)
+router.put('/:id/update', authenticateToken, async (req, res, next) => {
+  try {
+    const review = await getReviewById(req.params.id);
+    if (review.user_id !== req.user.id) {
+      return res.status(403).json({ message: 'You are not authorized to update this review' });
+    }
     const response = await updateReview(req.params.id, req.body.review);
     res.send(response);
   } catch (error) {
